Move Redux and Services to libs
