+++ apps/metin2/src/app/app.routing.module.ts
@@ -5,7 +5,7 @@ import { Routes, RouterModule, ExtraOptions } from '@angular/router';
 import { MainComponent } from './main/main.component';
 
 // Auth Guard
-import { AuthGuardService } from './services/auth/auth-guard.service';
+import { AuthGuardService } from '@metin2/api';
 
 
 // Components
