+++ apps/metin2/src/app/components/shared/login/login.component.ts
@@ -2,8 +2,8 @@ import { Component, OnInit } from '@angular/core';
 
 //Redux
 import { Store } from '@ngrx/store';
-import { AppState } from '../../../store/app.reducers';
-import { HiddenLoginModal } from '../../../store/actions';
+import { AppState } from '@store';
+// import { HiddenLoginModal } from '../../../store/actions';
 
 
 @Component({
