+++ apps/metin2/src/app/components/shared/user-manager/password/password.component.ts
@@ -1,10 +1,8 @@
 import { Component, OnInit } from '@angular/core';
 import { FormGroup, FormControl, Validators } from '@angular/forms';
 // Services
-import { HttpService } from '../../../../services/http/http.service';
+import { HttpService } from '@metin2/api';
 // Store
-import { Store } from '@ngrx/store';
-import { AppState } from '../../../../store/app.reducers';
 
 @Component({
   selector: 'manager-passwd',
