+++ apps/metin2/src/app/components/shared/user-manager/players/players.component.ts
@@ -2,7 +2,7 @@ import { Component, OnInit } from '@angular/core';
 
 // Redux
 import { Store } from '@ngrx/store';
-import { AppState } from '../../../../store/app.reducers';
+import { AppState } from '@store';
 import { Player } from '@metin2/api';
 
 @Component({
@@ -21,7 +21,7 @@ export class PlayersListComponent implements OnInit {
   ngOnInit() {
     this.store.select('user').subscribe(({players}) => {
       this.players = players;
-    })
+    });
   }
 
 }
