+++ apps/metin2/src/app/components/shared/user-manager/user-manager/user-manager.component.ts
@@ -2,14 +2,14 @@ import { Component, OnInit} from '@angular/core';
 import { Router } from '@angular/router';
 
 // Services
-import { AuthService } from '../../../../services/auth/auth.service';
+import { AuthService } from '@metin2/api';
 
 // Redux
 import { Store } from '@ngrx/store';
-import { AppState } from '../../../../store/app.reducers';
+import { AppState } from '@store';
 import {
   HiddenProfileModal,
-} from '../../../../store/actions';
+} from '@store/actions';
 
 
 @Component({
