+++ apps/metin2/src/app/components/shared/wigets/players/players.component.ts
@@ -2,8 +2,8 @@ import { Component, OnInit } from '@angular/core';
 
 // Store
 import { Store } from '@ngrx/store'
-import { AppState } from '../../../../store/app.reducers'
-import { ShowRankingPlayerModal } from '../../../../store/actions';
+import { AppState } from '@store'
+import { ShowRankingPlayerModal } from '@store/actions';
 
 @Component({
   selector: 'app-players',
