+++ libs/api/src/index.ts
@@ -1,1 +1,3 @@
 export * from './lib/shared.interface';
+export * from './lib/services/index';
+export { ServicesModule } from './lib/services/services.module';
