+++ libs/api/src/lib/services/http/authenticate.service.ts
