+++ libs/api/src/lib/services/index.ts
@@ -1,0 +1,15 @@
+// SocketService
+export { WebsocketService } from './websocket.service';
+
+// export services
+export { HttpService } from './http/http.service';
+export { AuthenticateService } from './http/authenticate.service';
+
+// Auth Services
+export { AuthService } from './auth/auth.service';
+export { AuthGuardService } from './auth/auth-guard.service';
+
+export { SignupService } from './signup.service';
+export { ErrorInterceptor } from './error-interceptor';
+
+export { token_getter } from './token-getter';
