+++ libs/redux/src/lib/index.ts
@@ -1,0 +1,1 @@
+export * from './store/app.reducers';
