+++ libs/redux/tsconfig.json
@@ -1,0 +1,7 @@
+{
+    "extends": "../../tsconfig.json",
+    "compilerOptions": {
+      "types": ["node"]
+    },
+    "include": ["**/*.ts"]
+  }
